+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Contacts/InternalContacts.cs
@@ -1,0 +1,6 @@
+﻿namespace DfE.FindInformationAcademiesTrusts.Data.Repositories.Contacts;
+
+public record InternalContacts(
+    InternalContact? TrustRelationshipManager,
+    InternalContact? SfsoLead
+);