+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Trust/TrustContactUpdated.cs
@@ -1,0 +1,5 @@
+﻿namespace DfE.FindInformationAcademiesTrusts.Data.Repositories.Trust;
+
+public record TrustContactUpdated(
+    bool EmailUpdated,
+    bool NameUpdated);