+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Trust/TrustContacts.cs
@@ -1,8 +1,6 @@
 ﻿namespace DfE.FindInformationAcademiesTrusts.Data.Repositories.Trust;
 
 public record TrustContacts(
-    Person? TrustRelationshipManager,
-    Person? SfsoLead,
     Person? AccountingOfficer,
     Person? ChairOfTrustees,
     Person? ChiefFinancialOfficer