+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Contacts/EditTrustRelationshipManager.cshtml.cs
@@ -1,8 +1,8 @@
 using DfE.FindInformationAcademiesTrusts.Configuration;
 using DfE.FindInformationAcademiesTrusts.Data;
+using DfE.FindInformationAcademiesTrusts.Data.Enums;
 using DfE.FindInformationAcademiesTrusts.Services.DataSource;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.FeatureManagement.Mvc;
 
 namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts.Contacts;
@@ -14,19 +14,10 @@ public class EditTrustRelationshipManagerModel(
     ILogger<EditTrustRelationshipManagerModel> logger,
     ITrustService trustService)
     : EditContactModel(trustProvider, dataSourceService, trustService,
-        logger, "Trust relationship manager")
+        logger, ContactRole.TrustRelationshipManager)
 {
-    public override async Task<IActionResult> OnGetAsync()
+    protected override InternalContact? GetContactFromServiceModel(TrustContactsServiceModel contacts)
     {
-        var pageResult = await base.OnGetAsync();
-
-        if (pageResult.GetType() == typeof(NotFoundResult)) return pageResult;
-
-        var contacts = await TrustService.GetTrustContactsAsync(Uid);
-
-        Email = contacts.TrustRelationshipManager?.Email;
-        Name = contacts.TrustRelationshipManager?.FullName;
-
-        return pageResult;
+        return contacts.TrustRelationshipManager;
     }
 }